53-FE: Move `flattenPages` into `shared/utils/pagination`
