+++ Okane.Client/src/features/financeRecords/components/FinanceRecordList.vue
@@ -7,12 +7,12 @@ import DeleteFinanceRecordModal from '@features/financeRecords/components/Delete
 import FinanceRecordListItem from '@features/financeRecords/components/FinanceRecordListItem.vue'
 import InfiniteScroller from '@shared/components/InfiniteScroller.vue'
 
-import { useInfiniteQueryFinanceRecords } from '@features/financeRecords/composables/useInfiniteQueryFinanceRecords'
-
-import { flattenPages } from '@shared/utils/response'
-
 import { FINANCES_COPY } from '@features/financeRecords/constants/copy'
 import { FINANCE_RECORD_SEARCH_FILTERS_KEY } from '@features/financeRecords/constants/searchFilters'
+
+import { useInfiniteQueryFinanceRecords } from '@features/financeRecords/composables/useInfiniteQueryFinanceRecords'
+
+import { flattenPages } from '@shared/utils/pagination'
 
 const searchFilters = inject(FINANCE_RECORD_SEARCH_FILTERS_KEY)
 const queryResult = useInfiniteQueryFinanceRecords(searchFilters)
