+++ Okane.Client/src/shared/utils/pagination.ts
@@ -2,6 +2,10 @@
 import type { InfiniteData } from '@tanstack/vue-query'
 import type { APIPaginatedResponse } from '@shared/services/apiClient/types'
 
+export function flattenPages<T>(pages?: APIPaginatedResponse<T>[]) {
+  return pages?.flatMap((page) => page.items) ?? []
+}
+
 export function removeItemFromPage<TItem>(
   data: InfiniteData<APIPaginatedResponse<TItem>>,
   shouldIncludeItem: (item: TItem) => boolean,
