--- Okane.Client/src/shared/utils/response.ts
@@ -1,6 +1,0 @@
-// Internal
-import type { APIPaginatedResponse } from '@shared/services/apiClient/types'
-
-export function flattenPages<T>(pages?: APIPaginatedResponse<T>[]) {
-  return pages?.flatMap((page) => page.items) ?? []
-}
